Fix #524 and #526: Broken in non-English Thunderbird
A string added in the Firefox 60 fix wasn't copied to the non-English `messages.json` files, which meant that it wasn't in the non-English `strings.properties` and `string.dtd` files. This caused Thunderbird to not load the overlay in Thunderbird, which prevented it from being present in the message window.

To help mitigate this in the future, to some degree, the `i18n.js` script makes sure that all English strings are present in the other languages when creating the the .properties and .dtd files. But I still have to remember to run that script.